+++ AsBuiltReport.Veeam.VBR/Src/Private/Report/Get-AbrVbrBackupjob.ps1
@@ -29,7 +29,7 @@ function Get-AbrVbrBackupjob {
     process {
         try {
             if ($Bkjobs = Get-VBRJob -WarningAction SilentlyContinue | Where-Object { $_.TypeToString -ne 'Windows Agent Backup' -and $_.TypeToString -ne 'Hyper-V Replication' -and $_.TypeToString -ne 'VMware Replication' } | Sort-Object -Property Name) {
-                $Bkjobs += [Veeam.Backup.Core.CBackupJob]::GetAll() | Where-Object { $_.TypeToString -like '*Nutanix*' } | Sort-Object -Property 'Name'
+                $Bkjobs += try { [Veeam.Backup.Core.CBackupJob]::GetAll() | Where-Object { $_.TypeToString -like '*Nutanix*' } | Sort-Object -Property 'Name' } catch { Out-Null }
                 $OutObj = @()
                 foreach ($Bkjob in $Bkjobs) {
                     try {